pyproject.toml
[project]
name = "vibesync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "langchain>=0.3.8",
    "langchain-community>=0.3.8",
    "msclap>=1.3.3",
]
